Add initial prototype
The first prototype of Te. This is a functional modal text editor for
the terminal, but only barely. We can open a file, move the cursor
around, insert some text, delete some text and then save our edits.

The only dependency is on `crossterm`, which means the rendering code is
implemented (not particularly well) in here. That took me a little
while. And we just use a `String` to hold the text. I tried to implement
a gap buffer at first, but that was taking me a bit longer than it
should have.

The next steps I'll probably follow are switching out to use `tui.rs`
and `ropey`, so someone else can handle the hard parts for me :P